+++ extractor-sdk/indexify_extractor_sdk/server_if/coordinator_service_pb2_grpc.py
